requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
numpy==1.26.0
Pillow==10.0.1
segment-anything==1.0
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78